requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
torch
